fix: tests fix
